// interfaces/lightning/LND/LND-channel.interface.ts
export interface LNDChannelInterface {
    capacity: number;
    commit_transaction_fee: number;
    commit_transaction_weight: number;
    id: string;
    is_active: boolean;
    is_closing: boolean;
    is_opening: boolean;
    is_partner_initiated: boolean;
    is_private: boolean;
    local_balance: number;
    local_csv: number;
    local_dust: number;
    local_given: number;
    local_max_htlcs: number;
    local_max_pending_mtokens: string;
    local_min_htlc_mtokens: string;
    local_reserve: number;
    partner_public_key: string;
    past_states: number;
    pending_payments: any[];
    received: number;
    remote_balance: number;
    remote_csv: number;
    remote_dust: number;
    remote_given: number;
    remote_max_htlcs: number;
    remote_max_pending_mtokens: string;
    remote_min_htlc_mtokens: string;
    remote_reserve: number;
    sent: number;
    time_offline: number;
    time_online: number;
    transaction_id: string;
    transaction_vout: number;
    unsettled_balance: number;
}